requirements.txt
SQLAlchemy~=2.0.27
python-dotenv~=1.0.1
fastapi~=0.112.2
pydantic~=2.1.1
requests~=2.31.0
urllib3~=2.2.1
celery~=5.4.0